[mod]:ui优化细节优化15.

+++ app/src/main/java/com/chocolate/nigerialoanapp/bean/response/HistoryRecordResponse.kt
@@ -5,7 +5,7 @@ class HistoryRecordResponse {
     var order_history: List<History>? = null  //列表对象
 
     inner class History {
-        var order_id: Int? = null  //订单ID
+        var order_id: Long? = null  //订单ID
         var overdue_day: Int? = null  //逾期天数
         var apply_date: String? = null  //申请日期
         var loan_amount: Int? = null  //借款金额
